Add interfaces for all processors
