+++ src/Hcal/HcalDoubleEndRecProducer.cxx
@@ -1,42 +1,6 @@
-#include "Conditions/SimpleTableCondition.h"
-#include "DetDescr/DetectorID.h"
-#include "DetDescr/HcalDigiID.h"
-#include "DetDescr/HcalGeometry.h"
-#include "DetDescr/HcalID.h"
-#include "Framework/EventDef.h"
-#include "Framework/EventProcessor.h"
-#include "Hcal/HcalReconConditions.h"
-#include "Recon/Event/HgcrocDigiCollection.h"
+#include "Hcal/HcalDoubleEndRecProducer.h"
 
 namespace hcal {
-
-class HcalDoubleEndRecProducer : public framework::Producer {
-  /// name of pass of rechits to use
-  std::string pass_name_{""};
-  /// name of rechits to use as input
-  std::string coll_name_{"HcalRecHits"};
-  /// name of pass of rechits to reconstruct
-  std::string rec_pass_name_{""};
-  /// name of rechits to reconstruct
-  std::string rec_coll_name_{"HcalRecHitsDoubleEnd"};
-
-  /// number of PEs per MIP
-  double pe_per_mip_;
-  /// energy per MIP [MeV]
-  double mip_energy_;
-  /// length of clock cycle [ns]
-  double clock_cycle_;
-
- private:
-
- public:
-  HcalDoubleEndRecProducer(const std::string& n, framework::Process& p)
-      : Producer(n, p) {}
-
-  virtual void configure(framework::config::Parameters& p) final override;
-  virtual void produce(framework::Event& event) final override;
-
-};  // HcalDoubleEndRecProducer
 
 void HcalDoubleEndRecProducer::configure(framework::config::Parameters& p) {
   pass_name_ = p.getParameter("pass_name", pass_name_);
@@ -57,8 +21,7 @@ void HcalDoubleEndRecProducer::produce(framework::Event& event) {
   const auto& conditions{
       getCondition<HcalReconConditions>(HcalReconConditions::CONDITIONS_NAME)};
 
-  auto hcalRecHits =
-      event.getCollection<ldmx::HcalHit>(coll_name_, pass_name_);
+  auto hcalRecHits = event.getCollection<ldmx::HcalHit>(coll_name_, pass_name_);
 
   std::vector<ldmx::HcalHit> doubleHcalRecHits;
 
@@ -78,21 +41,22 @@ void HcalDoubleEndRecProducer::produce(framework::Event& event) {
   // make pairs of hcal rechits indices that belong to the same pulse
   // @TODO: for now we just take the first two indices that have opposite-ends
   //        we do not cover the case where two hits come separated in time
-  std::map<ldmx::HcalID, std::pair<int,int>> indicesByID;
+  std::map<ldmx::HcalID, std::pair<int, int>> indicesByID;
   for (auto const& hcalBar : hitsByID) {
     auto id = hcalBar.first;
 
-    std::pair<int,int> indices(-1,-1);
+    std::pair<int, int> indices(-1, -1);
     int iHit = 0;
-    while ( iHit < hcalBar.second.size() ) {
+    while (iHit < hcalBar.second.size()) {
       auto hit = hcalBar.second.at(iHit);
 
-      ldmx::HcalDigiID digi_id(hit.getSection(), hit.getLayer(), hit.getStrip(), hit.getEnd());
-      if(digi_id.isNegativeEnd() && indices.second==-1) {
-	indices.second = iHit;
+      ldmx::HcalDigiID digi_id(hit.getSection(), hit.getLayer(), hit.getStrip(),
+                               hit.getEnd());
+      if (digi_id.isNegativeEnd() && indices.second == -1) {
+        indices.second = iHit;
       }
-      if(!digi_id.isNegativeEnd() && indices.first==-1) {
-	indices.first = iHit;
+      if (!digi_id.isNegativeEnd() && indices.first == -1) {
+        indices.first = iHit;
       }
       iHit++;
     }
@@ -108,58 +72,64 @@ void HcalDoubleEndRecProducer::produce(framework::Event& event) {
     const auto orientation{hcalGeometry.getScintillatorOrientation(id)};
 
     // skip non-double-ended layers
-    if (id.section() != ldmx::HcalID::HcalSection::BACK)
-      continue;
+    if (id.section() != ldmx::HcalID::HcalSection::BACK) continue;
 
     // get two hits to reconstruct
     auto hitPosEnd = hcalBar.second.at(indicesByID[id].first);
     auto hitNegEnd = hcalBar.second.at(indicesByID[id].second);
 
     // update TOA hit with negative end with mean shift
-    ldmx::HcalDigiID digi_id_pos(hitPosEnd.getSection(), hitPosEnd.getLayer(), hitPosEnd.getStrip(), hitPosEnd.getEnd());
-    ldmx::HcalDigiID digi_id_neg(hitNegEnd.getSection(), hitNegEnd.getLayer(), hitNegEnd.getStrip(), hitNegEnd.getEnd());
+    ldmx::HcalDigiID digi_id_pos(hitPosEnd.getSection(), hitPosEnd.getLayer(),
+                                 hitPosEnd.getStrip(), hitPosEnd.getEnd());
+    ldmx::HcalDigiID digi_id_neg(hitNegEnd.getSection(), hitNegEnd.getLayer(),
+                                 hitNegEnd.getStrip(), hitNegEnd.getEnd());
     double mean_shift = conditions.toaCalib(digi_id_neg.raw(), 1);
 
     double pos_time = hitPosEnd.getTime();
     double neg_time = hitNegEnd.getTime();
-    if( pos_time != 0 || neg_time !=0 ) {
+    if (pos_time != 0 || neg_time != 0) {
       neg_time = neg_time - mean_shift;
     }
 
     // update position in strip according to time measurement
     double v =
-      299.792 / 1.6;  // velocity of light in polystyrene, n = 1.6 = c/v
+        299.792 / 1.6;  // velocity of light in polystyrene, n = 1.6 = c/v
     double hitTimeDiff = pos_time - neg_time;
 
-    //std::cout << "\n new hit " << std::endl;
-    //std::cout << "strip " << id.strip() << " layer " << id.layer() << " center position " << position.X() << " " << position.Y() << " " << position.Z() << std::endl;
-    //std::cout << "hittime pos " << pos_time << " neg " << neg_time << " bar sign " << " diff " << hitTimeDiff << std::endl;
+    // std::cout << "\n new hit " << std::endl;
+    // std::cout << "strip " << id.strip() << " layer " << id.layer() << "
+    // center position " << position.X() << " " << position.Y() << " " <<
+    // position.Z() << std::endl; std::cout << "hittime pos " << pos_time << "
+    // neg " << neg_time << " bar sign " << " diff " << hitTimeDiff <<
+    // std::endl;
 
     int position_bar_sign = hitTimeDiff > 0 ? 1 : -1;
-    double position_unchanged = 0; int isX=0;
-    double position_bar =
-      position_bar_sign * fabs(hitTimeDiff) * v / 2;
-    if (orientation == 
-	ldmx::HcalGeometry::ScintillatorOrientation::horizontal) {
+    double position_unchanged = 0;
+    int isX = 0;
+    double position_bar = position_bar_sign * fabs(hitTimeDiff) * v / 2;
+    if (orientation ==
+        ldmx::HcalGeometry::ScintillatorOrientation::horizontal) {
       position_unchanged = position.X();
-      isX=1;
+      isX = 1;
       position.SetX(position_bar);
     } else {
       position_unchanged = position.Y();
-      isX=0;
+      isX = 0;
       position.SetY(position_bar);
     }
-    //std::cout << "position unchanged " << position_unchanged << " isx " << isX << std::endl;
-    //std::cout << "newposition " << position.X() << " " << position.Y() << " " << position.Z() << std::endl;
+    // std::cout << "position unchanged " << position_unchanged << " isx " <<
+    // isX << std::endl; std::cout << "newposition " << position.X() << " " <<
+    // position.Y() << " " << position.Z() << std::endl;
 
     // TODO: switch unique hit time for this pulse
     double hitTime = (hitPosEnd.getTime() + hitNegEnd.getTime());
 
     // amplitude and PEs
-    double num_mips_equivalent = (hitPosEnd.getAmplitude() + hitNegEnd.getAmplitude());
+    double num_mips_equivalent =
+        (hitPosEnd.getAmplitude() + hitNegEnd.getAmplitude());
     double PEs = (hitPosEnd.getPE() + hitNegEnd.getPE());
     double reconstructed_energy =
-      num_mips_equivalent * pe_per_mip_ * mip_energy_;
+        num_mips_equivalent * pe_per_mip_ * mip_energy_;
 
     // reconstructed Hit
     ldmx::HcalHit recHit;
@@ -171,7 +141,7 @@ void HcalDoubleEndRecProducer::produce(framework::Event& event) {
     recHit.setStrip(id.strip());
     recHit.setLayer(id.layer());
     recHit.setPE(PEs);
-    recHit.setMinPE(std::min(hitPosEnd.getPE(),hitNegEnd.getPE()));
+    recHit.setMinPE(std::min(hitPosEnd.getPE(), hitNegEnd.getPE()));
     recHit.setAmplitude(num_mips_equivalent);
     recHit.setAmplitudePos(hitPosEnd.getAmplitude());
     recHit.setAmplitudeNeg(hitNegEnd.getAmplitude());
